add database for locations and users as well as a little bit of UI for capekiwanda
