--- ZAMN/Controllers/CapeKiwanda.cs
@@ -1,16 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using System;
-
-namespace ZAMN.Controllers
-{
-  public class CapeKiwandaController : Controller
-  {
-
-    [HttpGet("/capeKiwanda")]
-    public ActionResult Index()
-    {
-      return View();
-    }
-
-  }
-}
