+++ ZAMN/Views/CapeKiwanda/New.cshtml
@@ -1,0 +1,19 @@
+@{
+  Layout = "_Layout";
+}
+
+<body>
+  <h1 class="welcome">Add a comment</h1>
+
+
+  <form action="/capeKiwanda" method="post">
+    <label for="userName" class="userName">Enter your name:</label><br>
+    <input id="userName" name="userName" type="text">
+
+    <label for="userComment" class="userComment">Enter your comment:</label><br>
+    <input id="userComment" name="userComment" type="text">
+
+    <button type="submit" class="btn btn-info btn-sm">+</button>
+  </form>
+
+</body>
